Added table prefix option to Drupal migration
It's good practice in the Drupal community to always prefix the tables with something so that if you ever need to host two sites in the same database then you will easily know which tables belongs to which site.

This commit adds an option to the Drupal migration scripts that makes it possible to add such a prefix to the migration query.