Update site dir with contents of #583 at rev cb19367.
This is the big site redesign by @cobyism. To work with the new
site deployment mechanism, it is being copied from the gh-pages
branch of that PR into the existing site dir. Before we finally
deploy the new site, we should remember to merge #583 so that we
keep all the history from that PR.