Moving Quick-start guide to own section
As a new user to jekyll I wanted to get up and running as quickly as possible.

The first thing I did was jump to the docs (missing the example on the homepage - my bad) and look for the quick start guide. Since I couldn't find one I went for the 'Basic Usage' section. I ended up muddling around for about 10 minutes until I hit the home page again to see the simple - and very easy - example of how to get started.

So, in this pull request I've moved the Quick-start Guide out to it's own section so that users who really are impatient (like me) can clearly look for and jump to the quick start guide, follow the super-easy steps, and be up and running in a matter of minutes.

Interested to hear if this is thought of as being a good idea.