Referencing the documentation for jekyll-import installation
Currently, there are only beta releases of the jekyll-import gem, so to
install it, you have to do `gem install jekyll-import --pre`. Instead
of giving instructions that don't work right now, reference the
documentation. That way, this doesn't need to be updated again when
jekyll-import has a non-beta version; just the documentation will need
updating.