Reverts 18c033dc5c00b2ce0b3528cb310bf3988377e39f
The merge that is reverted assigned categories to posts
based on the subfolders in the _posts directory and 
was merged under the understanding that it was fixing a bug.

Subfolders in the _posts directory should not assign 
metadata information to posts at this point in time and 
was not a bug.