Updated plugins: added jekyll-auto-image generator
Added to the plugin list: jekyll-auto-image. A generator that makes available the first image of a post.

By installing the plugin you will be able to access the first image with {{ @page.image }}. This plugin is useful to Include an image on your list of posts or to set a twitter card for each post/page.