Fix #4856: Don't rescue LoadError or bundler load errors for Bundler.
When it comes to bundler it's smart enough to know what to require, and in casees it's not, it's smart enough to accept :require.  In most cases when bundler has a LoadError (or otherwise) it's because there is a problem inside of the Gem itself and when this happens, Jekyll will happily let that error slip when it shouldn't, resulting in a badly placed error that is actually wrong. This corrects that so errors can surface properly.